fix: Always overwrite with tables, even if empty
While the last commit restored behavior to mostly where it was in
v0.15.2, this commit introduces a behavior change with the goal of
treating empty tables the same as non-empty ones (in particular,
`int_to_empty` is now treated the same as `int_to_non_empty`, both of
them overwriting the integer with the table).

Treating them the same makes a lot of sense logically, and the fact that
we weren't doing so is probably a bug in the old implementation. But it
is a notable behavior change, and one worth considering carefully.